requirements.txt
flask==3.0.3
werkzeug==3.0.4
pymupdf==1.24.10
pinecone-client==5.0.1
google-generativeai==0.7.2